Fix editor not showing tag dropdown
