+++ resources/js/components/card.js
@@ -1,5 +1,6 @@
-import {updateTag, createTag} from './tag.js'
+import {updateTag} from './tag.js'
 import {getCardContainer} from './container.js'
+import { createCard } from '../ui/card.js';
 
 //Note utility function
 export function getNewNoteId(currentNoteData, newNoteData) {
@@ -98,26 +99,6 @@ export function refreshCardContainerLayout() {
     let mediaItemContainer = getCardContainer();
     $(mediaItemContainer).masonry('reloadItems');
     $(mediaItemContainer).masonry('layout');
-}
-
-export function createCard(title, content, id) {
-
-    const cardHtml = `
-    <div class="card note" data-revision-count="1"  data-id="${id}" >
-        
-        <div class="card-body text-start">
-            <h5 class="card-title">${title}</h5>
-            <div class="card-content">${content}</div>
-        </div>
-    
-        <div class="card-tags d-flex flex-row flex-wrap ms-2"></div>
-    </div>
-    `
-
-    const card = document.createElement("div");
-    card.innerHTML = cardHtml;
-
-    return card;
 }
 
 //Update card logic
