+++ resources/js/components/editor.js
@@ -1,6 +1,7 @@
 import {getEditor, getEditPostEditor} from '../wysiwyg/tiptap.js'
 import { updatePage } from './container';
-import { createTag } from './tag.js';
+import { createTag } from '../ui/tag.js';
+import {createEditorTagCheckBox} from '../ui/tag.js';
 
 export function initClickSimpleEditor() {
     document.getElementById("simpleEditor").addEventListener('click', () => {
@@ -107,31 +108,7 @@ export function initClickRemoveEditorBtn() {
 }
 
 export function initClickEditorTagList() {
-    $("#editorTagList").on('click', function (e) {
-
-        function createEditorTagCheckBox(name, id) {
-            let tag = `
-            <label class="dropdown-item"> <input type="checkbox" name="${name}" value="${name}">
-                <span class="ps-2">${name}</span>
-            </label>
-            `;
-    
-            let labelTag = document.createElement("label");
-            labelTag.classList.add("dropdown-item", "tag-checkbox");
-            let inputTag = document.createElement("input");
-            inputTag.type = "checkbox";
-            inputTag.name = name;
-            inputTag.value = id;
-            let textTag = document.createElement("span");
-            textTag.classList.add("ps-2");
-            textTag.textContent = name;
-    
-            labelTag.append(inputTag);
-            labelTag.append(textTag);
-    
-            return labelTag;
-        }
-    
+    $("#tagNoteBtn").on('click', function (e) {
         $.ajax({
             url: "/load-tag",
             type: 'GET',
@@ -146,7 +123,6 @@ export function initClickEditorTagList() {
             }
         })
     });
-    
 }
 
 export function initHiddenEditor() {
