+++ resources/views/components/editor/add.blade.php
@@ -7,7 +7,7 @@
     <x-slot:dropdownTagListId>d</x-slot>
     <x-slot:tagsContainerId>f</x-slot>
     <x-slot:tagShowBtnId>d</x-slot>
-    <x-slot:dropdownTagListId>f</x-slot>
+    <x-slot:dropdownTagListCheckboxId>f</x-slot>
     <x-slot:submitTagBtnId>g</x-slot>
     <x-slot:buttons>
         <button class='btn btn-light border border-0' id='cancelBtn'><i class='bi bi-x'></i></button>
